pyproject.toml
[project]
name = "ai-agent-deployment-engineer-takehome"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.6.0",
    "openai-agents>=0.4.0",
    "python-dotenv>=1.1.1",
    "inquirer>=3.2.0",
    "halo>=0.0.31",
]
